requirements.txt
fastapi>=0.104.1
pydantic>=2.5.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
openai>=1.3.5
pinecone>=2.2.4
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-text-splitters>=1.0.0
langchain-community>=0.4.0
pypdf>=3.17.0

